client/requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-ahtx0==1.0.25
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
paho-mqtt==2.1.0
psutil==7.0.0
pyftdi==0.56.0
pyserial==3.5
python-dotenv==1.1.0
pyusb==1.3.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
sysv_ipc==1.1.0
typing_extensions==4.13.2